+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* 02/17/2024
+    * Fix multisampled texture preview
+
 * 02/16/2024
     * Fix inspecting async render and compute pipelines
     * Auto load textures in inspector